chore(composeables): renamed composeables
