+++ packages/composables/src/index.ts
@@ -12,10 +12,10 @@ export { default as useConfig } from './composables/useConfig';
 export { default as useFacet } from './composables/useFacet';
 export { default as useMakeOrder } from './composables/useMakeOrder';
 export { default as useMenuCategory } from './composables/useMenuCategory';
-export { default as usePage } from './composables/usePage';
+export { default as usePage } from './composables/useContent';
 export { default as useProduct } from './composables/useProduct';
 export { default as useReview } from './composables/useReview';
-export { default as useRouter } from './composables/useRouter';
+export { default as useRouter } from './composables/useUrlResolver';
 export { default as useShipping } from './composables/useShipping';
 export { default as useShippingProvider } from './composables/useShippingProvider';
 export { default as useUser } from './composables/useUser';
